Omi - OpenMW Mod Installer
Omi is a simple tool meant to ease the installation of OpenMW modifications. It's able to take a folder, determine whether or not that folder has the modification files, find and copty the modification files, and write an openmw.cfg entry.

Platforms supported:  Linux and other Unix/Unix-like systems (but not MacOS.)

To use omi, clone this repository and run __init__.py. You must have Python 3 and PyQt5 installed.